+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/features/sets/RelationValueAdapter.kt
@@ -214,8 +214,8 @@ class RelationValueAdapter(
         class Object(view: View) : ViewHolder(view), DragAndDropViewHolder {
             fun bind(item: RelationValueView.Object) = with(itemView) {
                 tvTitle.text = item.name
-                if (item.type != null) {
-                    tvSubtitle.text = item.type
+                if (item.typeName != null) {
+                    tvSubtitle.text = item.typeName
                 } else {
                     tvSubtitle.setText(R.string.unknown_object_type)
                 }